Add email notification functionality for person limit events and refactor email handling
